+++ client/src/components/BidsGraphAll.vue
